don't require numpy to run, unless using timeparse
Right now, every user of this django app must have numpy installed.
But numpy is currently only being used in utils.timeparse.
utils.timeparse is only being used by BigDateOGRFieldConverter,
which is only used by handlers.BigDateFieldConverterHandler.
The docstring for that class states that using it requires
a special build of GeoTools.
Unless that special build is a requirement of django-osgeo-importer,
any installation without it will already be unable to use
BigDateFieldConverterHandler, which is the only user of numpy.

Deferring this import allows individual installations to either use
or not use numpy depending on whether they also have the right
GeoTools setup and intend to use BigDateOGRFieldConverter.