set upload name and file type where feasible
Since some of the multifile and geopackage changes, it is no longer obvious
what should be the name of an upload. If it just contains one file, we can do
as before, but otherwise things can become more difficult.

The heuristics for this are a little complex, so they've been split off into
their own method and tests created just for this method.