add layer.layer_name
to api, UploadLayer, set in form processing, migration to add column.

layer.name keeps filename info that is used to import, so we can't stuff actual
layer names (e.g. from inside gpkg) there without breaking import. If we change
both, any existing databases will be storing a different kind of information in
that field, which may be confusing. We can deprecate layer.name if we need to.