don't use mutable objects as kwarg default values
What Python does here isn't what some would expect: it creates one mutable
object (list, dict, etc.) around import time, and then any time a value is not
passed for that argument, it uses that one mutable object. This makes it very
easy for the function body to mutate the argument and have this magically apply
to all other calls of the same function. It makes the function unexpectedly
stateful.

The usual pattern to avoid this is to use a sentinel value like None, which
cannot be mutated.